코드 정리

+++ List/CircularDoublyLinkedList.h
@@ -14,9 +14,9 @@ typedef struct tagCDLLNode
 CDLLNode* CDLL_CreateNode(ElementType NewData);
 void CDLL_DestroyNode(CDLLNode* Node);
 void CDLL_AppendNode(CDLLNode** Head, ElementType NewData);
-unsigned int CDLL_GetNodeCount(CDLLNode* Head);
-CDLLNode* CDLL_GetNodeAt(CDLLNode* Head, int Location);
-void CDLL_InsertAfter(CDLLNode* Head, int Location, ElementType NewData);
-void CDLL_RemoveNode(CDLLNode** Head, int Location);
+size_t CDLL_GetNodeCount(CDLLNode* Head);
+CDLLNode* CDLL_GetNodeAt(CDLLNode* Head, size_t Location);
+void CDLL_InsertAfter(CDLLNode* Head, size_t Location, ElementType NewData);
+void CDLL_RemoveNode(CDLLNode** Head, size_t Location);
 
 #endif
