+++ List/DoublyLinkedList.h
@@ -15,8 +15,8 @@ typedef struct tagDLLNode
 DLLNode* DLL_CreateNode(ElementType NewData);
 void DLL_DestroyNode(DLLNode* Node);
 void DLL_AppendNode(DLLNode** Head, ElementType NewData);
-unsigned int DLL_GetNodeCount(DLLNode* Head);
-DLLNode* DLL_GetNodeAt(DLLNode* Head, int Location);
-void DLL_RemoveNode(DLLNode** Head, int Location);
-void DLL_InsertAfter(DLLNode* Head, int Location, ElementType NewData);
+size_t DLL_GetNodeCount(DLLNode* Head);
+DLLNode* DLL_GetNodeAt(DLLNode* Head, size_t Location);
+void DLL_RemoveNode(DLLNode** Head, size_t Location);
+void DLL_InsertAfter(DLLNode* Head, size_t Location, ElementType NewData);
 #endif
