+++ List/LinkedList.h
@@ -13,13 +13,13 @@ typedef struct tagSLLNode
 SLLNode* SLL_CreateNode(ElementType NewData);
 void SLL_DestroyNode(SLLNode* Node);
 void SLL_AppendNode(SLLNode** Head, ElementType NewData);
-SLLNode* SLL_GetNodeAt(SLLNode* Head, int Location);
-void SLL_RemoveNode(SLLNode** Head, int Location);
-void SLL_InsertBefore(SLLNode** Head, int Location, ElementType NewData);
-void SLL_InsertAfter(SLLNode* Head, int Location, ElementType NewData);
+SLLNode* SLL_GetNodeAt(SLLNode* Head, size_t Location);
+void SLL_RemoveNode(SLLNode** Head, size_t Location);
+void SLL_InsertBefore(SLLNode** Head, size_t Location, ElementType NewData);
+void SLL_InsertAfter(SLLNode* Head, size_t Location, ElementType NewData);
 void SLL_InsertNewHead(SLLNode** Head, ElementType NewData);
 void SLL_DestroyAllNodes(SLLNode** Head);
-int SLL_GetNodeCount(SLLNode* Head);
+size_t SLL_GetNodeCount(SLLNode* Head);
 
 
 #endif
